Fix attempting to open screens on newly placed parts
It was *not* in fact a LMP issue.

Returning `ActionResult.PASS` when placing a block, even on the client, causes Minecraft to then attempt to use the block at the position being right-clicked. When replacing glass with a tank part, the grass would get replaced with the tank, but the client would think that nothing had happened, then attempting to use the newly-placed tank part block.